+++ skcriteria/preprocessing/distance.py
@@ -32,7 +32,9 @@ _skc_prep_scalers = "skcriteria.preprocessing.scalers"
 
 
 @deprecated(
-    reason=f"Use '{_skc_prep_scalers}.matrix_scale_by_cenit_distance' instead",
+    reason=(
+        f"Use ``{_skc_prep_scalers}.matrix_scale_by_cenit_distance`` instead"
+    ),
     version=0.8,
 )
 @doc_inherit(scalers.matrix_scale_by_cenit_distance)
@@ -41,7 +43,7 @@ def cenit_distance(matrix, objectives):
 
 
 @deprecated(
-    reason=f"Use '{_skc_prep_scalers}.CenitDistanceMatrixScaler' instead",
+    reason=f"Use ``{_skc_prep_scalers}.CenitDistanceMatrixScaler`` instead",
     version=0.8,
 )
 @doc_inherit(scalers.CenitDistanceMatrixScaler, warn_class=False)