+++ skcriteria/preprocessing/weighters.py
@@ -261,7 +261,7 @@ class EntropyWeighter(SKCWeighterABC):
 
 
 @deprecated(
-    reason="Please use pd.DataFrame(arr.T).correlation('pearson')",
+    reason="Please use ``pd.DataFrame(arr.T).correlation('pearson')``",
     version=0.8,
 )
 def pearson_correlation(arr):
@@ -291,7 +291,7 @@ def pearson_correlation(arr):
 
 
 @deprecated(
-    reason="Please use pd.DataFrame(arr.T).correlation('spearman')",
+    reason="Please use ``pd.DataFrame(arr.T).correlation('spearman')``",
     version=0.8,
 )
 def spearman_correlation(arr):
@@ -414,7 +414,7 @@ class CRITIC(SKCWeighterABC):
 
 
 @deprecated(
-    reason="Use 'skcriteria.preprocessing.weighters.CRITIC' instead",
+    reason="Use ``skcriteria.preprocessing.weighters.CRITIC`` instead",
     version=0.8,
 )
 @doc_inherit(CRITIC, warn_class=False)