+++ skcriteria/utils/doctools.py
@@ -20,8 +20,6 @@ from inspect import isclass
 
 from custom_inherit import doc_inherit as _doc_inherit
 
-from deprecated import deprecated as _deprecated
-
 # =============================================================================
 # DOC INHERITANCE
 # =============================================================================
@@ -64,48 +62,3 @@ def doc_inherit(parent, warn_class=True):
         return dec(obj)
 
     return _wrapper
-
-
-# =============================================================================
-# Deprecation
-# =============================================================================
-
-
-class SKCriteriaDeprecationWarning(DeprecationWarning):
-    """Skcriteria deprecation warning."""
-
-
-# _ If the version of the warning is >= ERROR_GE the action is setted to
-# 'error', otherwise is 'once'.
-ERROR_GE = 1.0
-
-
-def deprecated(*, reason, version):
-    """Mark functions, classes and methods as deprecated.
-
-    It will result in a warning being emitted when the object is called,
-    and the "deprecated" directive was added to the docstring.
-
-    Parameters
-    ----------
-    reason: str
-        Reason message which documents the deprecation in your library.
-    version: str
-        Version of your project which deprecates this feature.
-        If you follow the `Semantic Versioning <https://semver.org/>`_,
-        the version number has the format "MAJOR.MINOR.PATCH".
-
-    Notes
-    -----
-    This decorator is a thin layer over
-    :py:func:`deprecated.deprecated`.
-
-    Check: <github `https://pypi.org/project/Deprecated/`>__
-
-    """
-    return _deprecated(
-        reason=reason,
-        version=version,
-        category=SKCriteriaDeprecationWarning,
-        action=("error" if version >= ERROR_GE else "once"),
-    )