+++ tests/utils/test_doctools.py
@@ -25,8 +25,7 @@ import numpy as np
 
 import pytest
 
-from skcriteria.utils import decorators
-
+from skcriteria.utils import doctools
 
 # =============================================================================
 # TEST CLASSES
@@ -45,11 +44,11 @@ def test_doc_inherit():
 
     func_a.__doc__ = doc
 
-    @decorators.doc_inherit(func_a)
+    @doctools.doc_inherit(func_a)
     def func_b():
         ...
 
-    @decorators.doc_inherit(doc)
+    @doctools.doc_inherit(doc)
     def func_c():
         ...
 
@@ -58,49 +57,13 @@ def test_doc_inherit():
     # test warnings
     with pytest.warns(UserWarning):
 
-        @decorators.doc_inherit(doc, warn_class=True)
+        @doctools.doc_inherit(doc, warn_class=True)
         class A:  # noqa
             pass
 
     with warnings.catch_warnings():
         warnings.simplefilter("error")
 
-        @decorators.doc_inherit(doc, warn_class=False)
+        @doctools.doc_inherit(doc, warn_class=False)
         class A:  # noqa
             pass
-
-
-def test_deprecated():
-    def func():
-        """Zaraza.
-
-        Foo
-
-        Parameters
-        ----------
-        a: int
-            coso.
-
-        Returns
-        -------
-        None:
-            Nothing to return.
-
-        """
-        pass
-
-    expected_doc = func.__doc__
-
-    decorator = decorators.deprecated(reason="because foo", version=0.66)
-    func = decorator(func)
-
-    with pytest.deprecated_call():
-        func()
-
-    assert func.__doc__ == expected_doc
-
-    # this can be useful to catch bugs
-    # print("-" * 100)
-    # print(repr(func.__doc__))
-    # print(repr(expected_doc))
-    # print("-" * 100)